refactor: 라우트그룹 폴더 삭제 및 게시글 등록 디자인 수정
