--- src/app/(marketplace)/components/index.ts
@@ -1,8 +1,0 @@
-export { DetailHeader } from '@/app/(marketplace)/components/DetailHeader';
-export { DetailContent } from '@/app/(marketplace)/components/DetailContent';
-export { DetailContentFooter } from '@/app/(marketplace)/components/DetailContentFooter';
-export { CommentSection } from '@/app/(marketplace)/components/comment/CommentSection';
-export { DetailAside } from '@/app/(marketplace)/components/DetailAside/DetailAside';
-export { ShoppingMeetingRegisterModal } from '@/app/(marketplace)/components/registerModal/MeetingRegisterModal';
-export { ActionMenu } from '@/app/(marketplace)/components/ActionMenu/ActionMenu';
-export { IntroSection } from '@/app/(marketplace)/components/IntroSection';
