--- src/app/(marketplace)/layout.tsx
@@ -1,14 +1,0 @@
-import { SideButtonSection } from '@/components';
-
-export default function ShoppingLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      {children}
-      <SideButtonSection />
-    </div>
-  );
-}
