+++ src/app/shopping/[id]/page.tsx
