+++ src/app/shopping/components/ShoppingListSection.tsx
