+++ src/app/shopping/components/index.ts
