+++ src/components/marketplace/DetailContent.tsx
