+++ src/components/marketplace/DetailContentFooter.tsx
