+++ src/components/marketplace/index.ts
@@ -1,0 +1,8 @@
+export { DetailHeader } from '@/components/marketplace/DetailHeader';
+export { DetailContent } from '@/components/marketplace/DetailContent';
+export { DetailContentFooter } from '@/components/marketplace/DetailContentFooter';
+export { CommentSection } from '@/components/marketplace/comment/CommentSection';
+export { DetailAside } from '@/components/marketplace/DetailAside/DetailAside';
+export { ShoppingMeetingRegisterModal } from '@/components/marketplace/registerModal/MeetingRegisterModal';
+export { ActionMenu } from '@/components/marketplace/ActionMenu/ActionMenu';
+export { IntroSection } from '@/components/marketplace/IntroSection';
